(PC-18892)[PRO] refactor: rename CollectiveBookingDetails
