+++ pro/src/screens/Bookings/BookingsRecapTable/components/Table/Body/OldCollectiveBookingDetails/CollectiveBookingDetails.module.scss
