+++ pro/src/screens/Bookings/BookingsRecapTable/components/Table/Body/OldCollectiveBookingDetails/CollectiveBookingDetails.stories.tsx
