+++ pro/src/screens/Bookings/BookingsRecapTable/components/Table/Body/OldCollectiveBookingDetails/__specs__/CollectiveBookingDetails.spec.tsx
