+++ pro/src/screens/Bookings/BookingsRecapTable/components/Table/Body/OldCollectiveBookingDetails/index.ts
